Market_basket_analysis/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
pandas==1.3.3
mlxtend==0.19.0
matplotlib==3.4.3
